doubling

+++ CourseEvals.java
@@ -73,8 +73,8 @@ public class CourseEvals {
 
         rowScanner.close();
         System.out.println("Evaluation Done!");
-        int summeravg = (totalsummerscore / summercount);
-        int regularavg = (totalregularscore / regularcount);
+        double summeravg = (totalsummerscore / summercount);
+        double regularavg = (totalregularscore / regularcount);
         System.out.println("The average evaluation of a summer class is: " + summeravg);
         System.out.println("The average evaluation of a regular class is: " + regularavg);
